+++ server/src/controllers/map.controller.ts
@@ -1,10 +1,7 @@
 import { Controller, Get, Param, Query, Post, Body, UseGuards } from '@nestjs/common';
-import { ShortestPathService } from '../providers/shortest-path.service';
-import { MapUpdaterService } from '../providers/map-updater.service';
+import { ShortestPathService } from '../services/shortest-path.service';
+import { MapUpdaterService } from '../services/map-updater.service';
 import { AuthGuard } from '@nestjs/passport';
-
-// jest.setTimeout(-1);
-// jest.useFakeTimers();
 
 @Controller('map')
 @UseGuards(AuthGuard('jwt'))
