+++ server/src/services/srtm.service.ts
